Implement fast initialization of RNG states on the GPU.
By using different polynomials, we can update the xoroshiro generator as
if we'd performed 2, 4, 8, ... jumps. This in allows us to perform
arbitrarily sized jumps using log2(n) smaller jumps.

This is used to initialize the rng states directly on the GPU and
concurrently, which is much faster than initializing them sequentially
and then copying the state onto the CPU.